refactor code, so state manager is not the dependency any more
