--- src/stateManager/types.ts
@@ -1,6 +1,0 @@
-import { Context, Effect } from 'effect';
-
-export interface IdGenerator {
-  next(type: 'workflow' | 'workItem'): Effect.Effect<never, never, string>;
-}
-export const IdGenerator = Context.Tag<IdGenerator>();
